+++ clients/algoliasearch-client-dart/packages/client_core/README.md
@@ -1,8 +1,15 @@
-# Algolia Client Core
+<p align="center">
+  <a href="https://www.algolia.com">
+    <img alt="Algolia for Kotlin" src="https://raw.githubusercontent.com/algolia/algoliasearch-client-common/master/banners/dart.png" >
+  </a>
 
-[![Pub](https://img.shields.io/pub/v/algolia_client_core.svg)](https://pub.dartlang.org/packages/algolia_client_core)
+  <p align="center">
+    <a href="https://pub.dartlang.org/packages/algolia_client_core"><img src="https://img.shields.io/pub/v/algolia_client_core.svg" alt="Latest version"></img></a>
+    <a href="https://pub.dev/packages/algolia_client_core/publisher"><img src="https://img.shields.io/pub/publisher/algolia_client_core.svg" alt="Licence"></img></a>
+  </p>
+</p>
 
-Algolia Client Core is a Dart package designed to streamline your interactions with the Algolia API. It encompasses all the necessary logic for configuring and executing HTTP requests to the Algolia API. The package also offers robust support for retry strategies and comprehensive exception handling to ensure your application remains stable and efficient.
+**Algolia Client Core** is a Dart package designed to streamline your interactions with the Algolia API. It encompasses all the necessary logic for configuring and executing HTTP requests to the Algolia API. The package also offers robust support for retry strategies and comprehensive exception handling to ensure your application remains stable and efficient.
 
 ## Installation
 
@@ -23,4 +30,4 @@ dart pub get
 
 ## License
 
-Algolia Client Core is distributed under the [MIT license](LICENSE).
+Algolia Dart API Client is an open-sourced software licensed under the [MIT license](LICENSE).
