--- clients/algoliasearch-client-dart/packages/client_search_lite/analysis_options.yaml
@@ -1,1 +1,0 @@
-include: package:lints/recommended.yaml
