--- clients/algoliasearch-client-dart/packages/client_search_lite/build.yaml
@@ -1,14 +1,0 @@
-targets:
-  $default:
-    builders:
-      json_serializable:
-        options:
-          any_map: false
-          checked: true
-          create_factory: true
-          create_to_json: true
-          disallow_unrecognized_keys: false
-          explicit_to_json: true
-          field_rename: none
-          ignore_unannotated: false
-          include_if_null: false
