+++ templates/dart/lib.mustache
@@ -1,10 +1,10 @@
 // {{pubDescription}}
-library {{pubName}};
+library {{#isAlgoliasearchClient}}algoliasearch_lite{{/isAlgoliasearchClient}}{{^isAlgoliasearchClient}}{{pubName}}{{/isAlgoliasearchClient}};
 
 export 'package:algolia_client_core/algolia_client_core.dart'
     show AgentSegment, ApiClient, CallType, ClientOptions, Host, RequestOptions;
-{{#apiInfo}}{{#apis}}export 'package:{{pubName}}/{{sourceFolder}}/{{apiPackage}}/{{classFilename}}.dart';
+{{#apiInfo}}{{#apis}}export '{{sourceFolder}}/{{apiPackage}}/{{classFilename}}.dart';
 {{/apis}}{{/apiInfo}}
-{{#models}}{{#model}}export 'package:{{pubName}}/{{sourceFolder}}/{{modelPackage}}/{{classFilename}}.dart';
+{{#models}}{{#model}}export '{{sourceFolder}}/{{modelPackage}}/{{classFilename}}.dart';
 {{/model}}{{/models}}
-export 'package:{{pubName}}/src/extension.dart';
+export 'src/extension.dart';
