+++ templates/dart/pubspec.mustache
@@ -13,9 +13,13 @@ environment:
   sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
-  algolia_client_core: ^0.1.0
+  algolia_client_core: ^0.1.0-alpha.1
   json_annotation: ^4.8.1
   collection: ^1.17.2
+  {{#isAlgoliasearchClient}}
+  algolia_search: ^{{searchVersion}}
+  algolia_insights: ^{{insightsVersion}}
+  {{/isAlgoliasearchClient}}
 
 dev_dependencies:
   build_runner: ^2.4.4
